+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.12.2
+
+* Fix for Visual Crossing and Weatherbit.io snowfall amounts.
+* Time library updates.
+* Improved weather alert formatting.
+
 ## 2.12.1
 
 * Documentation corrections.